Reimplement int test
* Rename demo notebook
* Reimplement RUNME
* Reimplement GHA
