--- .github/workflows/integration-test-aws-main.yml
@@ -1,43 +1,0 @@
-name: AWS integration test main
-
-on:
- workflow_dispatch:
- push:
-    branches:
-      - main
-
-jobs:
- run-databricks-notebook:
-   runs-on: ubuntu-latest
-   steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2
-      - name: Run a databricks notebook
-        uses: databricks/run-notebook@v0
-        with:
-          local-notebook-path: RUNME.py
-          git-branch: main
-          databricks-host: https://e2-demo-west.cloud.databricks.com
-          databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_AWS }}
-          libraries-json: >
-            [
-              { "cran": {"package": "fable"} },
-              { "cran": {"package": "fabletools"} },
-              { "cran": {"package": "urca"} },
-              { "cran": {"package": "feasts"} },
-              { "cran": {"package": "lazyeval"} },
-              { "cran": {"package": "tsibble"} },
-              { "cran": {"package": "distributional"} }
-            ]
-          new-cluster-json: >
-            {
-              "num_workers": 2,
-              "spark_version": "11.3.x-cpu-ml-scala2.12",
-              "node_type_id": "i3.xlarge",
-              "aws_attributes": {"first_on_demand": 1,"availability": "ON_DEMAND"}
-            }
-          notebook-params-json: >
-            {
-              "run_job": "True"
-            }
-
