+++ .github/workflows/integration-test-aws-pr.yml
@@ -7,34 +7,40 @@ jobs:
  run-databricks-notebook:
    runs-on: ubuntu-latest
    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2
-      - name: Run a databricks notebook
-        uses: databricks/run-notebook@v0
-        with:
-          local-notebook-path: RUNME.py
-          git-commit: ${{ github.event.pull_request.head.sha }}
-          databricks-host: https://e2-demo-west.cloud.databricks.com
-          databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_AWS }}
-          libraries-json: >
-            [
-              { "cran": {"package": "fable"} },
-              { "cran": {"package": "fabletools"} },
-              { "cran": {"package": "urca"} },
-              { "cran": {"package": "feasts"} },
-              { "cran": {"package": "lazyeval"} },
-              { "cran": {"package": "tsibble"} },
-              { "cran": {"package": "distributional"} }
-            ]
-          new-cluster-json: >
-            {
-              "num_workers": 2,
-              "spark_version": "11.3.x-cpu-ml-scala2.12",
-              "node_type_id": "i3.xlarge",
-              "aws_attributes": {"first_on_demand": 1,"availability": "ON_DEMAND"}
-            }
-          notebook-params-json: >
-            {
-              "run_job": "True"
-            }
-            
+     - name: Checkout repo
+       uses: actions/checkout@v2
+     - name: Run a databricks notebook
+       uses: databricks/run-notebook@v0
+       with:
+         local-notebook-path: RUNME.py
+         git-commit: ${{ github.event.pull_request.head.sha }}
+         databricks-host: https://e2-demo-west.cloud.databricks.com
+         databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_AWS }}
+         new-cluster-json: >
+           {
+             "num_workers": 0,
+             "spark_version": "10.4.x-scala2.12",
+             "node_type_id": "i3.xlarge",
+             "aws_attributes": {
+                 "availability": "ON_DEMAND"
+              },
+             "spark_conf": {
+                  "spark.master": "local[*, 4]",
+                  "spark.databricks.cluster.profile": "singleNode"
+              },
+              "custom_tags": {
+                  "ResourceClass": "SingleNode"
+              }
+           }
+         notebook-params-json: >
+           {
+             "run_job": "True"
+           }
+         access-control-list-json: >
+           [
+             {
+               "group_name": "users",
+               "permission_level": "CAN_VIEW"
+             }
+           ]
+
