+++ .github/workflows/integration-test-aws-push.yml
@@ -1,6 +1,11 @@
-name: MSA integration test PR
+name: AWS integration test push
+
 on:
- pull_request:
+ workflow_dispatch:
+ push:
+    branches:
+      - main
+      - web-sync
 
 jobs:
  run-databricks-notebook:
@@ -12,17 +17,17 @@ jobs:
        uses: databricks/run-notebook@v0
        with:
          local-notebook-path: RUNME.py
-         git-commit: ${{ github.event.pull_request.head.sha }}
-         databricks-host: https://adb-984752964297111.11.azuredatabricks.net
-         databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_MSA }}
+         git-commit: ${{ github.sha }}
+         databricks-host: https://e2-demo-west.cloud.databricks.com
+         databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_AWS }}
          new-cluster-json: >
            {
              "num_workers": 0,
              "spark_version": "10.4.x-scala2.12",
-             "node_type_id": "Standard_DS3_v2",
-             "azure_attributes": {
-                  "availability": "ON_DEMAND_AZURE"
-                            },
+             "node_type_id": "i3.xlarge",
+             "aws_attributes": {
+               "availability": "ON_DEMAND"
+             },
              "spark_conf": {
                   "spark.master": "local[*, 4]",
                   "spark.databricks.cluster.profile": "singleNode"
@@ -30,7 +35,6 @@ jobs:
               "custom_tags": {
                   "ResourceClass": "SingleNode"
               }
-                            
            }
          notebook-params-json: >
            {
