--- .github/workflows/integration-test-gcp-main.yml
@@ -1,44 +1,0 @@
-name: GCP integration test main
-
-on:
-  workflow_dispatch:
-  push:
-    branches:
-      - main
-
-jobs:
-  run-databricks-notebook:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout repo
-        uses: actions/checkout@v2
-      - name: Run a databricks notebook
-        uses: databricks/run-notebook@v0
-        with:
-          local-notebook-path: RUNME.py
-          git-branch: main
-          databricks-host: https://416411475796958.8.gcp.databricks.com
-          databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_GCP }}
-          libraries-json: >
-            [
-              { "cran": {"package": "fable"} },
-              { "cran": {"package": "fabletools"} },
-              { "cran": {"package": "urca"} },
-              { "cran": {"package": "feasts"} },
-              { "cran": {"package": "lazyeval"} },
-              { "cran": {"package": "tsibble"} },
-              { "cran": {"package": "distributional"} }
-            ]
-          new-cluster-json: >
-            {
-              "num_workers": 2,
-              "spark_version": "11.3.x-cpu-ml-scala2.12",
-              "node_type_id": "n1-highmem-4",
-              "gcp_attributes": {
-                "availability": "ON_DEMAND_GCP"
-              }
-            }
-          notebook-params-json: >
-            {
-              "run_job": "True"
-            }
