+++ .github/workflows/integration-test-gcp-pr.yml
@@ -16,26 +16,30 @@ jobs:
          git-commit: ${{ github.event.pull_request.head.sha }}
          databricks-host: https://416411475796958.8.gcp.databricks.com
          databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_GCP }}
-         libraries-json: >
-           [
-             { "cran": {"package": "fable"} },
-             { "cran": {"package": "fabletools"} },
-             { "cran": {"package": "urca"} },
-             { "cran": {"package": "feasts"} },
-             { "cran": {"package": "lazyeval"} },
-             { "cran": {"package": "tsibble"} },
-             { "cran": {"package": "distributional"} }
-           ]
          new-cluster-json: >
            {
-             "num_workers": 2,
-             "spark_version": "11.3.x-cpu-ml-scala2.12",
-             "node_type_id": "n1-highmem-4",
-             "gcp_attributes": {
-               "availability": "ON_DEMAND_GCP"
-             }
-           }
+              "num_workers": 0,
+              "spark_version": "10.4.x-scala2.12",
+              "node_type_id": "n1-highmem-4",
+              "gcp_attributes": {
+                "availability": "ON_DEMAND_GCP"
+              },
+              "spark_conf": {
+                  "spark.master": "local[*, 4]",
+                  "spark.databricks.cluster.profile": "singleNode"
+              },
+              "custom_tags": {
+                  "ResourceClass": "SingleNode"
+              }
+            }
          notebook-params-json: >
            {
-             "run_job": "True"
+            "run_job": "True"
            }
+         access-control-list-json: >
+           [
+             {
+               "group_name": "users",
+               "permission_level": "CAN_VIEW"
+             }
+           ]
