--- .github/workflows/integration-test-msa-main.yml
@@ -1,44 +1,0 @@
-name: MSA integration test main
-on:
- workflow_dispatch:
- push:
-    branches:
-      - main
-
-jobs:
- run-databricks-notebook:
-   runs-on: ubuntu-latest
-   steps:
-     - name: Checkout repo
-       uses: actions/checkout@v2
-     - name: Run a databricks notebook
-       uses: databricks/run-notebook@v0
-       with:
-         local-notebook-path: RUNME.py
-         git-branch: main
-         databricks-host: https://adb-984752964297111.11.azuredatabricks.net
-         databricks-token: ${{ secrets.DEPLOYMENT_TARGET_TOKEN_MSA }}
-         libraries-json: >
-           [
-             { "cran": {"package": "fable"} },
-             { "cran": {"package": "fabletools"} },
-             { "cran": {"package": "urca"} },
-             { "cran": {"package": "feasts"} },
-             { "cran": {"package": "lazyeval"} },
-             { "cran": {"package": "tsibble"} },
-             { "cran": {"package": "distributional"} }
-           ]
-         new-cluster-json: >
-           {
-             "num_workers": 2,
-             "spark_version": "11.3.x-cpu-ml-scala2.12",
-             "node_type_id": "Standard_DS3_v2",
-             "azure_attributes": {
-                "availability": "ON_DEMAND_AZURE",
-                "first_on_demand": 1
-                }
-           }
-         notebook-params-json: >
-           {
-             "run_job": "True"
-           }
